Keep (TCNT0-lasttime) to 8-bits, as it'll never be more than that anyway.

+++ clockless_trinket.h
@@ -260,7 +260,7 @@ protected:
 				hi = *port | mask;
 				lo = *port & ~mask;
 
-				if((TCNT0-lasttime) > ((WAIT_TIME-INTERRUPT_THRESHOLD)/US_PER_TICK)) { sei(); return; }
+				if((uint8_t)(TCNT0-lasttime) > ((WAIT_TIME-INTERRUPT_THRESHOLD)/US_PER_TICK)) { sei(); return; }
 				// Sum of the clock counts across each row should be 10 for 8Mhz, WS2811
 				// The values in the D1/D2/D3 indicate how many cycles the previous column takes
 				// to allow things to line back up.
